Rename metric.
The metric is now "numSplits" and its description (what shows up in the
Spark UI) is "number of file splits read".
Also changed "IcebergScan" to "IcebergBatchScan" in the `description()`
and `toString()` of `SparkBatchQueryScan`.
