+++ spark/v3.2/spark/src/main/java/org/apache/iceberg/spark/source/TaskNumSplits.java
@@ -19,27 +19,22 @@
 
 package org.apache.iceberg.spark.source;
 
-import java.text.NumberFormat;
-import org.apache.spark.sql.connector.metric.CustomMetric;
+import org.apache.spark.sql.connector.metric.CustomTaskMetric;
 
-public class NumFiles implements CustomMetric {
+public class TaskNumSplits implements CustomTaskMetric {
+  private long value;
+
+  TaskNumSplits(long value) {
+    this.value = value;
+  }
 
   @Override
   public String name() {
-    return "numFiles";
+    return "numSplits";
   }
 
   @Override
-  public String description() {
-    return "number of files read";
-  }
-
-  @Override
-  public String aggregateTaskMetrics(long[] taskMetrics) {
-    long sum = initialValue;
-    for (int i = 0; i < taskMetrics.length; i++) {
-      sum += taskMetrics[i];
-    }
-    return NumberFormat.getIntegerInstance().format(sum);
+  public long value() {
+    return value;
   }
 }
